expose min/max registration
